requirements.txt
fastapi
mangum
pandas
scikit-learn
sentence-transformers
line-bot-sdk
google-generativeai
googletrans==4.0.0-rc1
python-dotenv